+++ docs/changelog.md
@@ -1,5 +1,15 @@
 # Changelog
 
+## [v0.3.3](https://github.com/0x2142/frigate-notify/releases/tag/v0.3.3) - Aug 14 2024
+ - Additional filtering options for alerts:
+     - [Sublabels](https://frigate-notify.0x2142.com/latest/config/file/#sublabels)
+     - [Minimum label score](https://frigate-notify.0x2142.com/latest/config/file/#labels)
+     - [Quiet hours](https://frigate-notify.0x2142.com/latest/config/file/#quiet-hours)
+ - Ability to [modify snapshot](https://frigate-notify.0x2142.com/latest/config/file/#general) when retrieved from Frigate: include bounding box / timestamp or crop snapshot
+ - Allow use of [template variables](https://frigate-notify.0x2142.com/latest/config/templates/#available-variables) within HTTP headers sent to notification providers
+     - Ntfy action button can now be overrriden by defining a custom `X-Actions` header in the config file
+     - Templates can also now include information from [environment variables](https://frigate-notify.0x2142.com/latest/config/templates/#environment-variables)
+
 ## [v0.3.2](https://github.com/0x2142/frigate-notify/releases/tag/v0.3.2) - Jun 13 2024
 
  - Allow config to be provided via environment variables (Thanks [@ryan-willis](https://github.com/ryan-willis)!!)